+++ tests/testthat.R
@@ -1,9 +1,4 @@
-
 library(testthat)
-library(ggplot2)
-library(ggseg)
-library(ggseg3d)
+library(ggsegJHU)
 
 test_check("ggsegJHU")
-
-# covr::zero_coverage(covr::package_coverage("."))